fix: ignore test bots

+++ src/app/struct/Tracker.js
@@ -142,7 +142,9 @@ class Tracker {
 			} else {
 				Logger.warn(`Channel: ${clan.channel}`, { level: 'Missing Channel' });
 				this.delete(clan.guild, clan.tag);
-				await Clans.destroy({ where: { channel: clan.channel } });
+				if (this.client.user.id === process.env.CLIENT_ID) {
+					await Clans.destroy({ where: { channel: clan.channel } });
+				}
 			}
 
 			await this.delay(100);
